# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NovelHub",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/kb/entities": {
      "get": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "List Entities Endpoint",
        "operationId": "list_entities_endpoint_api_kb_entities_get",
        "parameters": [
          {
            "name": "novel_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          },
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Create Entity Endpoint",
        "operationId": "create_entity_endpoint_api_kb_entities_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/kb/entities/{entity_id}": {
      "get": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Get Entity Endpoint",
        "operationId": "get_entity_endpoint_api_kb_entities__entity_id__get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Entity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Update Entity Endpoint",
        "operationId": "update_entity_endpoint_api_kb_entities__entity_id__patch",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Entity Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Delete Entity Endpoint",
        "operationId": "delete_entity_endpoint_api_kb_entities__entity_id__delete",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Entity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Entity Endpoint Api Kb Entities  Entity Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/kb/resolve": {
      "post": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Resolve Alias Endpoint",
        "operationId": "resolve_alias_endpoint_api_kb_resolve_post",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "novel_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/EntityResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Resolve Alias Endpoint Api Kb Resolve Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/kb/merge": {
      "post": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Merge Entities Endpoint",
        "operationId": "merge_entities_endpoint_api_kb_merge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/kb/entities/{entity_id}/mentions": {
      "post": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Add Mention Endpoint",
        "operationId": "add_mention_endpoint_api_kb_entities__entity_id__mentions_post",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Entity Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MentionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Add Mention Endpoint Api Kb Entities  Entity Id  Mentions Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/graph/{novel_id}": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "Get Graph Endpoint",
        "operationId": "get_graph_endpoint_api_graph__novel_id__get",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          },
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Graph Endpoint Api Graph  Novel Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/graph/{novel_id}/path": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "Shortest Path Endpoint",
        "operationId": "shortest_path_endpoint_api_graph__novel_id__path_get",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          },
          {
            "name": "source_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Source Id"
            }
          },
          {
            "name": "target_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Target Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Shortest Path Endpoint Api Graph  Novel Id  Path Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events": {
      "post": {
        "tags": [
          "discovery"
        ],
        "summary": "Create Event Endpoint",
        "operationId": "create_event_endpoint_api_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{novel_id}": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "List Events Endpoint",
        "operationId": "list_events_endpoint_api_events__novel_id__get",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          },
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          },
          {
            "name": "chapter_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Chapter Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EntityCreate": {
        "properties": {
          "novel_id": {
            "type": "integer",
            "title": "Novel Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type",
            "default": "character"
          },
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes",
            "default": {}
          },
          "source_chapter": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Chapter"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aliases",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "novel_id",
          "name"
        ],
        "title": "EntityCreate"
      },
      "EntityList": {
        "properties": {
          "entities": {
            "items": {
              "$ref": "#/components/schemas/EntityResponse"
            },
            "type": "array",
            "title": "Entities"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "entities",
          "total"
        ],
        "title": "EntityList"
      },
      "EntityResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "novel_id": {
            "type": "integer",
            "title": "Novel Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes"
          },
          "source_chapter": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Chapter"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aliases"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "novel_id",
          "name",
          "entity_type",
          "attributes",
          "source_chapter",
          "aliases",
          "created_at",
          "updated_at"
        ],
        "title": "EntityResponse"
      },
      "EntityUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "entity_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Type"
          },
          "attributes": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attributes"
          }
        },
        "type": "object",
        "title": "EntityUpdate"
      },
      "EventCreate": {
        "properties": {
          "novel_id": {
            "type": "integer",
            "title": "Novel Id"
          },
          "chapter_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Id"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type",
            "default": "general"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "importance": {
            "type": "integer",
            "title": "Importance",
            "default": 1
          },
          "participants": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Participants",
            "default": []
          },
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "novel_id"
        ],
        "title": "EventCreate"
      },
      "EventList": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventResponse"
            },
            "type": "array",
            "title": "Events"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "events",
          "total"
        ],
        "title": "EventList"
      },
      "EventResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "novel_id": {
            "type": "integer",
            "title": "Novel Id"
          },
          "chapter_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Id"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "importance": {
            "type": "integer",
            "title": "Importance"
          },
          "participants": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Participants"
          },
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "novel_id",
          "chapter_id",
          "event_type",
          "title",
          "description",
          "importance",
          "participants",
          "attributes",
          "created_at"
        ],
        "title": "EventResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MentionCreate": {
        "properties": {
          "chapter_id": {
            "type": "integer",
            "title": "Chapter Id"
          },
          "position": {
            "type": "integer",
            "title": "Position",
            "default": 0
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "chapter_id"
        ],
        "title": "MentionCreate"
      },
      "MergeRequest": {
        "properties": {
          "source_entity_id": {
            "type": "integer",
            "title": "Source Entity Id"
          },
          "target_entity_id": {
            "type": "integer",
            "title": "Target Entity Id"
          },
          "keep_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keep Name"
          }
        },
        "type": "object",
        "required": [
          "source_entity_id",
          "target_entity_id"
        ],
        "title": "MergeRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}